Namespaces poo

+++ src/Controllers/PostsController.php
@@ -2,81 +2,108 @@
 
 namespace App\controllers;
 
-use Repository\PostRepository;
-use DatabaseConnection;
-use Entity\Post;
+use App\Model\Repository\PostRepository;
+use App\Model\Repository\CommentRepository;
+use App\Model\Repository\UserRepository;
+use App\lib\DatabaseConnection;
+use App\Model\Entity\Post;
 
 class PostsController
 {
     private $postRepository;
+    private $commentRepository;
+    private $userRepository;
     private $currentTime;
 
-    public function __construct(\DatabaseConnection $connection)
+    public function __construct(DatabaseConnection $connection)
     {
         $this->postRepository = new PostRepository($connection);
+        $this->commentRepository = new CommentRepository($connection);
+        $this->userRepository = new UserRepository($connection);
         $this->currentTime = date('Y-m-d H:i:s');
     }
 
 
     public function getPosts()
     {
-        $this->postRepository->getAllPosts();
-        require(__DIR__.'/../../public/templates/admin/admin_posts_list_page.php');
+        $posts = $this->postRepository->getAllPosts();
+        require_once(__DIR__.'/../../templates/admin/admin_posts_list_page.php');
     }
 
     public function getPublishedPosts()
     {
+
         $posts = $this->postRepository->getPublishedPosts();
-        require(__DIR__.'/../../public/templates/post_list_page.php');
+        require(__DIR__.'/../../templates/posts_list_page.php');
     }
 
     public function getPostById($postId)
     {
-        $this->postRepository->getPostById($postId);
+        $post = $this->postRepository->getPostById($postId);
+        $comments = $this->commentRepository->getValidComments($post->id);
+        require(__DIR__.'/../../templates/show_post_page.php');
     }
 
     public function getAddPost()
     {
-        require(__DIR__.'/../../public/templates/admin/posts/add_post_page.php');
+        $admins = $this->userRepository->getAdmins();
+        require(__DIR__.'/../../templates/admin/posts/add_post_page.php');
     }
 
-    public function postAddPost($user_id, $title, $chapo, $content, $is_published)
+    public function postAddPost($data)
     {
+        if (!isset($data["title"]) || !isset($data["content"])) {
+            echo "Veuillez remplir les champs Titre et Contenu.";
+            return;
+        }
+
+
+        $user = $this->userRepository->getUserById($data["user_id"]);
+
         $post = new Post();
-        $post->setTitle($title);
-        $post->setChapo($chapo);
-        $post->setContent($content);
-        $post->setAuthor($user_id);
-        $post->setIsPublished($is_published);
-        $post->setCreatedAt($this->currentTime);
-        $post->setUpdatedAt($this->currentTime);
+        $post->init(
+            $data["title"],
+            $data["chapo"],
+            $data["content"],
+            intval($data["user_id"]),
+            isset($data["is_published"]) ? 1 : 0
+        );
 
         $this->postRepository->addPost($post);
 
-        require('templates/admin/admin_posts_list_page.php');
+        require(__DIR__.'/../../templates/admin/posts/post_add_page.php');
     }
+
 
     public function getEditPost($postId)
     {
         $post = $this->postRepository->getPostById($postId);
-        require('templates/admin/posts/edit_post_page.php');
+        $admins = $this->userRepository->getAdmins();
+        require(__DIR__.'/../../templates/admin/posts/edit_post_page.php');
     }
 
-    public function postEditPost($postId, $user_id, $title, $chapo, $content, $is_published)
+    public function postEditPost($data)
     {
-        $post = $this->postRepository->getPostById($postId);
 
-        if($post) {
-            $post->setTitle($title);
-            $post->setChapo($chapo);
-            $post->setContent($content);
-            $post->setAuthor($user_id);
-            $post->setIsPublished($is_published);
-            $post->setUpdatedAt($this->currentTime);
+        if (!isset($data["post_id"]) || !isset($data["user_id"]) || !isset($data["title"]) || !isset($data["content"])) {
+            echo "Veuillez fournir toutes les informations nécessaires.";
+            return;
+        }
+
+        $post = $this->postRepository->getPostById($data["post_id"]);
+
+        if ($post) {
+            $post->update(
+                $data["title"],
+                $data["chapo"],
+                $data["content"],
+                $data["user_id"],
+                isset($data["is_published"]) ? 1 : 0
+            );
 
             $this->postRepository->editPost($post);
 
-            require('templates/admin/admin_posts_list_page.php');
+            require(__DIR__.'/../../templates/admin/admin_posts_list_page.php');
         } else {
             echo 'Ce post n\'existe pas';
         }
@@ -85,21 +112,24 @@ class PostsController
     public function getDeletedPost($postId)
     {
         $post = $this->postRepository->getPostById($postId);
-        require('templates/admin/posts/delete_post_page.php');
+        require(__DIR__.'/../../templates/admin/posts/delete_post_page.php');
     }
 
-    public function postDeletePost($postId)
+    public function postDeletePost($data)
     {
-        $post = $this->postRepository->getPostById($postId);
+        if (!isset($data["post_id"])) {
+            echo "Veuillez fournir l'identifiant du post.";
+            return;
+        }
+
+        $post = $this->postRepository->getPostById($data["post_id"]);
 
         if($post) {
-            $post->setIsPublished(0);
-            $post->setUpdatedAt($this->currentTime);
-            $post->setDeletedAt($this->currentTime);
+            $post->delete($data["post_id"], 0);
 
             $this->postRepository->deletePost($post);
 
-            require('templates/admin/admin_posts_list_page.php');
+            require(__DIR__.'/../../templates/admin/admin_posts_list_page.php');
 
         } else {
             echo 'Ce post n\'existe pas';
@@ -109,20 +139,19 @@ class PostsController
     public function getRestorePost($postId)
     {
         $post = $this->postRepository->getPostById($postId);
-        require('templates/admin/posts/restore_post_page.php');
+        require(__DIR__.'/../../templates/admin/posts/restore_post_page.php');
     }
 
-    public function restorePost($postId)
+    public function postRestorePost($data)
     {
-        $post = $this->postRepository->getPostById($postId);
+        $post = $this->postRepository->getPostById($data["post_id"]);
 
         if($post) {
-            $post->setDeletedAt(null);
-            $post->setUpdatedAt($this->currentTime);
+            $post->restore($data["post_id"]);
 
             $this->postRepository->restorePost($post);
 
-            require('templates/admin/admin_posts_list_page.php');
+            require(__DIR__.'/../../templates/admin/admin_posts_list_page.php');
         } else {
             echo 'Ce post n\'existe pas';
         }
